+++ firstSolar/back/src/routes/swap.ts
@@ -10,7 +10,7 @@ import { abi as DfsEthPoolAbi } from "../../contracts/artifacts/LiquidityPool.js
 
 // MainNet
 const web3 = new Web3(
-  "wss://polygon-mumbai.infura.io/ws/v3/2ca09ab04a7c44dcb6f886deeba97502"
+  "wss://polygon-mumbai.g.alchemy.com/v2/U60psLWRd8tg7yShqQgZ-1YTMSYB0EGo"
 );
 
 const router = express.Router();
@@ -22,28 +22,16 @@ router.post("/swapApprove", async (req, res) => {
     const filterToken = async () => {
       switch (req.body.data) {
         case "dfs":
-          return new web3.eth.Contract(
-            DFSAbi as AbiItem[],
-            process.env.DFS_TOKEN_CA
-          );
+          return new web3.eth.Contract(DFSAbi as AbiItem[], process.env.DFS);
 
         case "eth":
-          return new web3.eth.Contract(
-            DFSAbi as AbiItem[],
-            process.env.ETH_TOKEN_CA
-          );
+          return new web3.eth.Contract(DFSAbi as AbiItem[], process.env.ETH);
 
         case "usdt":
-          return new web3.eth.Contract(
-            DFSAbi as AbiItem[],
-            process.env.USDT_TOKEN_CA
-          );
+          return new web3.eth.Contract(DFSAbi as AbiItem[], process.env.USDT);
 
         case "bnb":
-          return new web3.eth.Contract(
-            DFSAbi as AbiItem[],
-            process.env.BNB_TOKEN_CA
-          );
+          return new web3.eth.Contract(DFSAbi as AbiItem[], process.env.BNB);
         default:
           throw new Error("Invalid token");
       }
@@ -78,17 +66,17 @@ router.post("/swapTransaction", async (req, res) => {
       if ("dfsethpool".includes(target)) {
         return new web3.eth.Contract(
           DfsEthPoolAbi as AbiItem[],
-          process.env.DFS_ETH_POOL_CA
+          process.env.DFS_ETH
         );
       } else if ("dfsbnbpool".includes(target)) {
         return new web3.eth.Contract(
           DfsEthPoolAbi as AbiItem[],
-          process.env.DFS_BNB_POOL_CA
+          process.env.DFS_BNB
         );
       } else if ("dfsusdtpool".includes(target)) {
         return new web3.eth.Contract(
           DfsEthPoolAbi as AbiItem[],
-          process.env.DFS_USDT_POOL_CA
+          process.env.DFS_USDT
         );
       } else {
         throw new Error("Invalid Error");
@@ -98,6 +86,7 @@ router.post("/swapTransaction", async (req, res) => {
     const amount = BigNumber.from(
       Math.floor(req.body.amount * 10 ** 18).toString()
     );
+    console.log("result.options", result.options.address);
 
     const tokenSwap = result.methods
       .swapTokens(result.options.address, amount)
