+++ firstSolar/back/src/routes/sync.ts
@@ -9,10 +9,11 @@ import { AbiItem } from "web3-utils";
 const web3 = new Web3(
   "wss://polygon-mumbai.g.alchemy.com/v2/U60psLWRd8tg7yShqQgZ-1YTMSYB0EGo"
 );
-import { abi as DFSAbi } from "../../contracts/artifacts/Token.json";
-import { abi as DfsEthPoolAbi } from "../../contracts/artifacts/LiquidityPool.json";
+
 import { abi as DexlAbi } from "../../contracts/artifacts/Dex.json";
+import price from "../priceList";
 import { type } from "os";
+import { send } from "process";
 
 const router = express.Router();
 router.use(express.json());
@@ -20,246 +21,23 @@ router.use(express.urlencoded({ extended: true }));
 
 router.get("/", async (req, res) => {
   try {
-    const SyncDB = async () => {
-      const DbList = await db.Price.findAll();
-      if (DbList.length <= 0) {
-        const BNBPrice = (
-          await axios.get(
-            "https://pro-api.coinmarketcap.com/v2/cryptocurrency/quotes/latest",
-            {
-              headers: {
-                "X-CMC_PRO_API_KEY": "95b65a08-8e34-4bcf-88f2-23f733797852",
-              },
-              params: {
-                symbol: "BNB",
-              },
-            }
-          )
-        ).data;
-
-        const ETHPrice = (
-          await axios.get(
-            "https://pro-api.coinmarketcap.com/v2/cryptocurrency/quotes/latest",
-            {
-              headers: {
-                "X-CMC_PRO_API_KEY": "95b65a08-8e34-4bcf-88f2-23f733797852",
-              },
-              params: {
-                symbol: "ETH",
-              },
-            }
-          )
-        ).data;
-
-        const USDTPrice = (
-          await axios.get(
-            "https://pro-api.coinmarketcap.com/v2/cryptocurrency/quotes/latest",
-            {
-              headers: {
-                "X-CMC_PRO_API_KEY": "95b65a08-8e34-4bcf-88f2-23f733797852",
-              },
-              params: {
-                symbol: "USDT",
-              },
-            }
-          )
-        ).data;
-
-        const bnbPriceToUSDT: Response = (
-          await axios.get(
-            "https://pro-api.coinmarketcap.com/v2/tools/price-conversion",
-            {
-              headers: {
-                "X-CMC_PRO_API_KEY": "95b65a08-8e34-4bcf-88f2-23f733797852",
-              },
-              params: {
-                amount: 10,
-                symbol: "BNB",
-                convert: "USDT",
-              },
-            }
-          )
-        ).data;
-
-        const bnbPriceToETH: Response = (
-          await axios.get(
-            "https://pro-api.coinmarketcap.com/v2/tools/price-conversion",
-            {
-              headers: {
-                "X-CMC_PRO_API_KEY": "95b65a08-8e34-4bcf-88f2-23f733797852",
-              },
-              params: {
-                amount: 10,
-                symbol: "BNB",
-                convert: "ETH",
-              },
-            }
-          )
-        ).data;
-
-        const ETHPriceToBNB: Response = (
-          await axios.get(
-            "https://pro-api.coinmarketcap.com/v2/tools/price-conversion",
-            {
-              headers: {
-                "X-CMC_PRO_API_KEY": "95b65a08-8e34-4bcf-88f2-23f733797852",
-              },
-              params: {
-                amount: 10,
-                symbol: "ETH",
-                convert: "BNB",
-              },
-            }
-          )
-        ).data;
-
-        const EthPriceToUSDT: Response = (
-          await axios.get(
-            "https://pro-api.coinmarketcap.com/v2/tools/price-conversion",
-            {
-              headers: {
-                "X-CMC_PRO_API_KEY": "95b65a08-8e34-4bcf-88f2-23f733797852",
-              },
-              params: {
-                amount: 10,
-                symbol: "ETH",
-                convert: "USDT",
-              },
-            }
-          )
-        ).data;
-
-        const usdtPriceToETH: Response = (
-          await axios.get(
-            "https://pro-api.coinmarketcap.com/v2/tools/price-conversion",
-            {
-              headers: {
-                "X-CMC_PRO_API_KEY": "95b65a08-8e34-4bcf-88f2-23f733797852",
-              },
-              params: {
-                amount: 10,
-                symbol: "USDT",
-                convert: "ETH",
-              },
-            }
-          )
-        ).data;
-
-        const usdtPriceToBNB: Response = (
-          await axios.get(
-            "https://pro-api.coinmarketcap.com/v2/tools/price-conversion",
-            {
-              headers: {
-                "X-CMC_PRO_API_KEY": "95b65a08-8e34-4bcf-88f2-23f733797852",
-              },
-              params: {
-                amount: 10,
-                symbol: "USDT",
-                convert: "BNB",
-              },
-            }
-          )
-        ).data;
-
-        // USDT
-
-        let USDT = Object.values(USDTPrice.data);
-        let USDTdata = USDT[0];
-        let USDTprice = Object.values(USDTdata[0]["quote"]);
-        let NowUSDTprice = USDTprice[0]["price"];
-
-        // USDT DB
-
-        let USDTIdData = USDTdata[0]["id"];
-        let USDTSymbolData = USDTdata[0]["symbol"];
-        let USDTSlugData = USDTdata[0]["slug"];
-
-        // ETH
-        let ETH = Object.values(ETHPrice.data);
-        let ETHdata = ETH[0];
-        let ETHprice = Object.values(ETHdata[0]["quote"]);
-        let NowETHprice = ETHprice[0]["price"];
-
-        // ETH DB
-        let ETHIdData = ETHdata[0]["id"];
-        let ETHSymbolData = ETHdata[0]["symbol"];
-        let ETHSlugData = ETHdata[0]["slug"];
-
-        // BNB
-        let BNB = Object.values(BNBPrice.data);
-        let BNBdata = BNB[0];
-        let BNBprice = Object.values(BNBdata[0]["quote"]);
-        let NowBNBprice = BNBprice[0]["price"];
-
-        // BNB DB
-        let BNBIdData = BNBdata[0]["id"];
-        let BNBSymbolData = BNBdata[0]["symbol"];
-        let BNBSlugData = BNBdata[0]["slug"];
-
-        // swap
-        const BNBSwapUSDT: any = Object.values(bnbPriceToUSDT)[1];
-        const BNBtoUSDT = BNBSwapUSDT[0].quote.USDT.price;
-
-        const BNBSwapETH: any = Object.values(bnbPriceToETH)[1];
-        const BNBtoETH = BNBSwapETH[0].quote.ETH.price;
-
-        const ETHSwapBNB: any = Object.values(ETHPriceToBNB)[1];
-        const ETHtoBNB = ETHSwapBNB[0].quote.BNB.price;
-
-        const ETHSwapUSDT: any = Object.values(EthPriceToUSDT)[1];
-        const ETHtoUSDT = ETHSwapUSDT[0].quote.USDT.price;
-
-        const USDTwapETH: any = Object.values(usdtPriceToETH)[1];
-        const USDTtoETH = USDTwapETH[0].quote.ETH.price;
-
-        const USDTwapBNB: any = Object.values(usdtPriceToBNB)[1];
-        const USDTtoBNB = USDTwapBNB[0].quote.BNB.price;
-
-        const DBConnet: any = [];
-        DBConnet.push({
-          IdData: USDTIdData,
-          symbolData: USDTSymbolData,
-          slugData: USDTSlugData,
-          convertUSDT: 1,
-          convertETH: USDTtoETH,
-          convertBNB: USDTtoBNB,
-          tokenPrice: NowUSDTprice,
+    const DbList = await db.Price.findAll();
+    if (DbList.length <= 0) {
+      let priceDataList = await price();
+      for (let i = 0; i < priceDataList.length; i++) {
+        await db.Price.create({
+          tokenSymbol: priceDataList[i].symbolData,
+          tokenId: priceDataList[i].IdData,
+          tokenSlug: priceDataList[i].slugData,
+          ConvertToUSDT: priceDataList[i].convertUSDT,
+          ConvertToETH: priceDataList[i].convertETH,
+          ConvertToBNB: priceDataList[i].convertBNB,
+          tokenPrice: priceDataList[i].tokenPrice,
         });
-        DBConnet.push({
-          IdData: ETHIdData,
-          symbolData: ETHSymbolData,
-          slugData: ETHSlugData,
-          convertUSDT: ETHtoUSDT,
-          convertETH: 1,
-          convertBNB: ETHtoBNB,
-          tokenPrice: NowETHprice,
-        });
-        DBConnet.push({
-          IdData: BNBIdData,
-          symbolData: BNBSymbolData,
-          slugData: BNBSlugData,
-          convertUSDT: BNBtoUSDT,
-          convertETH: BNBtoETH,
-          convertBNB: 1,
-          tokenPrice: NowBNBprice,
-        });
-        for (let i = 0; i < DBConnet.length; i++) {
-          await db.Price.create({
-            tokenSymbol: DBConnet[i].symbolData,
-            tokenId: DBConnet[i].IdData,
-            tokenSlug: DBConnet[i].slugData,
-            ConvertToUSDT: DBConnet[i].convertUSDT,
-            ConvertToETH: DBConnet[i].convertETH,
-            ConvertToBNB: DBConnet[i].convertBNB,
-            tokenPrice: DBConnet[i].tokenPrice,
-          });
-        }
       }
-    };
-    SyncDB();
+    }
 
     res.end();
-    console.log("insert Data OK");
   } catch (error) {
     console.log(error);
     res.end();
@@ -330,33 +108,130 @@ router.post("/mypage", async (req, res) => {
   }
 });
 
-router.post("/swap", (req, res) => {
-  const swapApp = async () => {
-    const BtnUSDT = await db.Price.findOne({
-      where: {
-        tokenSymbol: req.body.data,
-      },
-    });
-    res.send(BtnUSDT);
-  };
-  swapApp();
+router.post("/swap", async (req, res) => {
+  const BtnUSDT = await db.Price.findOne({
+    where: {
+      tokenSymbol: req.body.data,
+    },
+  });
+  res.send(BtnUSDT);
 });
-router.post("/filterSwap", (req, res) => {
-  const filterSwapApp = async () => {
-    console.log(req.body);
-    const BtnUSDT = await db.Price.findOne({
-      where: {
-        tokenSymbol: req.body.data,
-      },
-    });
-    res.send(BtnUSDT.tokenPrice);
-    // const BtnEth = await db.Price.findone({ tokenPrice });
-    // const BtnBnB = await db.Price.findone({});
-    // const swapEth = await db.Price.findone({
-    //   where: {},
-    // });
-  };
-  filterSwapApp();
+router.post("/filterSwap", async (req, res) => {
+  console.log(req.body);
+  const BtnUSDT = await db.Price.findOne({
+    where: {
+      tokenSymbol: req.body.data,
+    },
+  });
+  res.send(BtnUSDT.tokenPrice);
 });
+router.get("/dateSync", async (req, res) => {
+  try {
+    const timeSync = async () => {
+      console.log("실행됨 1");
+      const updateTime = await db.Price.findAll();
+      const nowTime = new Date();
 
+      updateTime.forEach(async (pool) => {
+        console.log("실행됨 2");
+
+        const lastUpdateTime = new Date(pool.dataValues.updatedAt);
+        // DB에서 불러온 각 Pool 인스턴스의 업데이트 시간을 가져와 'lastUpdateTime' 변수에 할당
+
+        const timeDiffInMs = nowTime.getTime() - lastUpdateTime.getTime();
+        // 현재 시간과 'lastUpdateTime'의 차이(밀리초)를 계산하여 'timeDiffInMs' 변수에 할당
+
+        const timeDiffInHours = timeDiffInMs / (1000 * 60 * 60);
+        // 'timeDiffInMs'를 시간 단위로 변환하여 'timeDiffInHours' 변수에 할당
+
+        // 'timeDiffInHours'가 24시간을 초과 하는지 여부에 따라 메세지 출력
+
+        const DbList = await db.Price.findAll();
+
+        if (DbList.length <= 0) {
+          console.log("실행됨 3");
+
+          // 24시간이 지났고, db에 값이 없을때
+          let priceDataList = await price();
+          for (let i = 0; i < priceDataList.length; i++) {
+            console.log("실행됨 4");
+
+            await db.Price.create({
+              tokenSymbol: priceDataList[i].symbolData,
+              tokenId: priceDataList[i].IdData,
+              tokenSlug: priceDataList[i].slugData,
+              ConvertToUSDT: priceDataList[i].convertUSDT,
+              ConvertToETH: priceDataList[i].convertETH,
+              ConvertToBNB: priceDataList[i].convertBNB,
+              tokenPrice: priceDataList[i].tokenPrice,
+            });
+          }
+        } else {
+          console.log("실행됨 5");
+
+          if (timeDiffInHours > 24) {
+            console.log("24시간 이상 지났습니다.");
+            let priceDataList = await price();
+            let priceDB = await db.Price.findAll();
+            for (let i = 0; i < priceDataList.length; i++) {
+              console.log("실행됨 6");
+
+              await db.Price.update(
+                {
+                  ConvertToUSDT: priceDB[i].convertUSDT,
+                  ConvertToETH: priceDB[i].convertETH,
+                  ConvertToBNB: priceDB[i].convertBNB,
+                  tokenPrice: priceDB[i].tokenPrice,
+                },
+                {
+                  where: {
+                    tokenSymbol: priceDataList[i].symbolData,
+                    tokenId: priceDataList[i].IdData,
+                    tokenSlug: priceDataList[i].slugData,
+                  },
+                }
+              );
+            }
+            priceDB = await db.Price.findAll();
+            console.log(priceDB);
+            return;
+          } else {
+            let priceDB = await db.Price.findAll();
+            // console.log("priceDB", priceDB);
+            // console.log("24시간 이내입니다.");
+            return;
+          }
+
+          // 24시간이 지났고, db에 지난 값들이 있을때
+        }
+      });
+    };
+    timeSync();
+    const now = new Date();
+
+    const nextUpdate = new Date(
+      now.getFullYear(),
+      now.getMonth(),
+      now.getDate() + 1,
+      9
+    );
+
+    // 다음 업데이트 시간까지 남은 시간(밀리초) 계산
+    const timeUntilNextUpdate = nextUpdate.getTime() - now.getTime();
+
+    // setInterval() 메서드로 주기적으로 checkUpdateTime() 함수 실행
+    setInterval(() => {
+      timeSync();
+    }, 60 * 1000);
+
+    // 첫 번째 업데이트는 다음 업데이트 시간까지 남은 시간 후에 실행
+    setTimeout(() => {
+      timeSync();
+    }, timeUntilNextUpdate);
+
+    res.end();
+  } catch (error) {
+    console.log(error);
+  }
+});
 export default router;
