+++ firstSolar/front/src/components/API/APICon.jsx
@@ -68,12 +68,7 @@ const APICon = () => {
   }, [volum, secondSelectTokenPrice, secondKeyWord, selectedSecondSwapToken]);
 
   useEffect(() => {
-    const firstSync = axios.get("http://localhost:8080/api/sync");
-    if (firstSync) {
-      return;
-    } else {
-      firstSync();
-    }
+    axios.get("http://localhost:8080/api/sync");
   }, []);
 
   useEffect(() => {
